Fix build

+++ include/game/board/player.h
@@ -43,7 +43,7 @@ void BoardPlayerPosGet(s32, Vec*);
 void BoardPlayerCoinsSet(s32 player, s32 value);
 s32 BoardPlayerSizeGet(s32 player);
 void BoardPlayerAutoSizeSet(s32 player, s32 value);
-void BoardPlayerMotionStart(s32, s32, s32);
+void BoardPlayerMotionStart(s16, s32, s32);
 void BoardPlayerMotionShiftSet(s32 player, s32 arg1, float start, float end, u32 attr);
 void BoardPlayerRotYSet(s32 player, float rot_y);
